+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Added Deck class (ordered object)
 * Added unit tests for Iterator & Deck
 * Added String#dissect(open, close)
+* Added RegExp.interpret(pattern)
 
 ## 0.1.1 (2014-06-23)